requirements.txt
requests
prefect
mysql-connector-python
konlpy